Conflicts:
	oc_framework-test-project/src/com/owncloud/android/oc_framework_test_project/TestActivity.java

+++ src/com/owncloud/android/datamodel/FileDataStorageManager.java
@@ -434,7 +434,7 @@ public class FileDataStorageManager {
                     }
                     updateFolderSize(file.getParentId());
                 }
-                if (removeLocalCopy && file.isDown()) {
+                if (removeLocalCopy && file.isDown() && file.getStoragePath() != null) {
                     boolean success = new File(file.getStoragePath()).delete();
                     if (!removeDBData && success) {
                         // maybe unnecessary, but should be checked TODO remove if unnecessary
